webapp: further simplifying App.tsx (moving utility code to utils/)
